move structs out to a new file.
