+++ internal/smash/app.go
@@ -23,41 +23,6 @@ import (
 	"github.com/thushan/smash/pkg/indexer"
 )
 
-type Flags struct {
-	Base           []string `yaml:"base"`
-	ExcludeDir     []string `yaml:"exclude-dir"`
-	ExcludeFile    []string `yaml:"exclude-file"`
-	Algorithm      int      `yaml:"algorithm"`
-	MaxThreads     int      `yaml:"max-threads"`
-	MaxWorkers     int      `yaml:"max-workers"`
-	DisableSlicing bool     `yaml:"disable-slicing"`
-	Silent         bool     `yaml:"silent"`
-	Verbose        bool     `yaml:"verbose"`
-}
-
-type App struct {
-	Flags     *Flags
-	Args      []string
-	Locations []string
-}
-type RunSummary struct {
-	DuplicateFileSizeF string
-	DuplicateFileSize  uint64
-	TotalFiles         int64
-	TotalFileErrors    int64
-	ElapsedTime        int64
-	UniqueFiles        int64
-	DuplicateFiles     int64
-}
-type SmashFile struct {
-	Filename    string
-	Hash        string
-	FileSizeF   string
-	FileSize    uint64
-	ElapsedTime int64
-	FullHash    bool
-}
-
 func (app *App) Run() error {
 
 	var locations = app.Locations
@@ -185,11 +150,12 @@ func (app *App) summariseSmashedFile(cache *haxmap.Map[string, []SmashFile], sta
 		FileSizeF:   humanize.Bytes(stats.FileSize),
 		ElapsedTime: ms,
 	}
-	if v, existing := cache.Get(sf.Hash); existing {
+	hash := sf.Hash
+	if v, existing := cache.Get(hash); existing {
 		v = append(v, sf)
-		cache.Set(sf.Hash, v)
+		cache.Set(hash, v)
 	} else {
-		cache.Set(sf.Hash, []SmashFile{sf})
+		cache.Set(hash, []SmashFile{sf})
 	}
 }
