+++ internal/smash/structs.go
@@ -1,0 +1,36 @@
+package smash
+
+type Flags struct {
+	Base           []string `yaml:"base"`
+	ExcludeDir     []string `yaml:"exclude-dir"`
+	ExcludeFile    []string `yaml:"exclude-file"`
+	Algorithm      int      `yaml:"algorithm"`
+	MaxThreads     int      `yaml:"max-threads"`
+	MaxWorkers     int      `yaml:"max-workers"`
+	DisableSlicing bool     `yaml:"disable-slicing"`
+	Silent         bool     `yaml:"silent"`
+	Verbose        bool     `yaml:"verbose"`
+}
+
+type App struct {
+	Flags     *Flags
+	Args      []string
+	Locations []string
+}
+type RunSummary struct {
+	DuplicateFileSizeF string
+	DuplicateFileSize  uint64
+	TotalFiles         int64
+	TotalFileErrors    int64
+	ElapsedTime        int64
+	UniqueFiles        int64
+	DuplicateFiles     int64
+}
+type SmashFile struct {
+	Filename    string
+	Hash        string
+	FileSizeF   string
+	FileSize    uint64
+	ElapsedTime int64
+	FullHash    bool
+}
